Move global object registries into their own file
