+++ yt/data_objects/api.py
@@ -1,6 +1,5 @@
 from . import construction_data_containers as __cdc, selection_data_containers as __sdc
 from .analyzer_objects import AnalysisTask, analysis_task
-from .data_containers import data_object_registry
 from .grid_patch import AMRGridPatch
 from .image_array import ImageArray
 from .octree_subset import OctreeSubset
