+++ yt/data_objects/derived_quantities.py
@@ -3,14 +3,13 @@ import numpy as np
 from yt.funcs import camelcase_to_underscore, ensure_list
 from yt.units.yt_array import array_like_field
 from yt.utilities.exceptions import YTParticleTypeNotFound
+from yt.utilities.object_registries import derived_quantity_registry
 from yt.utilities.parallel_tools.parallel_analysis_interface import (
     ParallelAnalysisInterface,
     parallel_objects,
 )
 from yt.utilities.physical_constants import gravitational_constant_cgs
 from yt.utilities.physical_ratios import HUGE
-
-derived_quantity_registry = {}
 
 
 def get_position_fields(field, data):
