+++ yt/utilities/object_registries.py
@@ -1,0 +1,7 @@
+# These are some of the data object registries that are used in different places in the code.  Not all of the self-registering objects are included in these.
+
+analysis_task_registry = {}
+data_object_registry = {}
+derived_quantity_registry = {}
+output_type_registry = {}
+simulation_time_series_registry = {}
