+++ yt/utilities/parameter_file_storage.py
@@ -4,12 +4,11 @@ from itertools import islice
 
 from yt.config import ytcfg
 from yt.funcs import mylog
+from yt.utilities.object_registries import output_type_registry
 from yt.utilities.parallel_tools.parallel_analysis_interface import (
     parallel_simple_proxy,
 )
 
-output_type_registry = {}
-simulation_time_series_registry = {}
 _field_names = ("hash", "bn", "fp", "tt", "ctid", "class_name", "last_seen")
