+++ yt/visualization/volume_rendering/old_camera.py
@@ -5,7 +5,6 @@ import numpy as np
 
 from yt.config import ytcfg
 from yt.data_objects.api import ImageArray
-from yt.data_objects.data_containers import data_object_registry
 from yt.funcs import ensure_numpy_array, get_num_threads, get_pbar, iterable, mylog
 from yt.units.yt_array import YTArray
 from yt.utilities.amr_kdtree.api import AMRKDTree
@@ -24,6 +23,7 @@ from yt.utilities.lib.image_samplers import (
 from yt.utilities.lib.misc_utilities import lines
 from yt.utilities.lib.partitioned_grid import PartitionedGrid
 from yt.utilities.math_utils import get_rotation_matrix
+from yt.utilities.object_registries import data_object_registry
 from yt.utilities.orientation import Orientation
 from yt.utilities.parallel_tools.parallel_analysis_interface import (
     ParallelAnalysisInterface,
